Moved malloc and tune to a specific targets directory, cudaStreamSynchronize wrapped in quda_cuda api
